refactor: modularize CLI commands
