+++ src/tests/test_cli_vault_stats.py
@@ -3,7 +3,7 @@ from types import SimpleNamespace
 from typer.testing import CliRunner
 
 from seedpass.cli import app
-from seedpass import cli
+from seedpass.cli import common as cli_common
 
 runner = CliRunner()
 
@@ -16,7 +16,7 @@ def test_vault_stats_command(monkeypatch):
     pm = SimpleNamespace(
         get_profile_stats=lambda: stats, select_fingerprint=lambda fp: None
     )
-    monkeypatch.setattr(cli, "PasswordManager", lambda: pm)
+    monkeypatch.setattr(cli_common, "PasswordManager", lambda: pm)
     result = runner.invoke(app, ["vault", "stats"])
     assert result.exit_code == 0
     out = result.stdout
